In this commit the: api, json, models, serializers, url basic setup where made.
For the Registrar project the aim is to aid students and future students with common questions they might have. The model for student has its rough start with the name, gps and address of the students. More fields maybe created later on when more help ideas aer created to be answer.  The api, json view and serializers are all setup to work on this basic build. The location model followed the same pattern. This is just the basic outline for the api, json, models, serializers, urls so that they made be further expanded on as this project is created.